feat: 새로운 하단바 생성, floating button 수정(삭제, 저장소로 이동, 뒤로가기)
