+++ lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:my_first_flutter_project/app_main_page.dart';
 import 'package:my_first_flutter_project/models/display.dart';
 import 'package:my_first_flutter_project/screens/0_archive_page.dart';
 import 'package:my_first_flutter_project/screens/1_home_page.dart';
@@ -26,6 +27,7 @@ class MyApp extends StatelessWidget{
       home: LoginPage(),
       routes: {
         '/login_page.dart': (context) => LoginPage(),
+        '/app_main_page.dart': (context) => MyAppPage(),
         '/1_home_page.dart': (context) => HomePage(),
         '/2_add_page.dart': (context) => AddPage(),
         '/0_archive_page.dart': (context) => ArchivePage(),
